% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datum2df_patch.cohort.R
\name{datum2df_patch.cohort}
\alias{datum2df_patch.cohort}
\title{datum2df_patch.cohort}
\usage{
datum2df_patch.cohort(datum, vars = c("agb.recr", "agb.mort"))
}
\arguments{
\item{datum}{datum}

\item{vars}{vars}

\item{patches}{pfts}
}
\value{
df
}
\description{
datum2df_patch.cohort
}
\author{
FM
}
