% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datum2df_patch.R
\name{datum2df_patch}
\alias{datum2df_patch}
\title{datum2df}
\usage{
datum2df_patch(
  datum,
  vars = c("gpp", "npp", "agb.change", "lai.change", "agb", "lai"),
  patches
)
}
\arguments{
\item{datum}{datum}

\item{vars}{vars}

\item{patches}{pfts}
}
\value{
df
}
\description{
datum2df_patch
}
\author{
FM
}
